🧐(project) add method to retrieve inactive user count from edx database
To be able to warn/delete inactive users by batch, add a method to get the
inactive users count.
As we now have multiple queries for getting inactive users informations, we
should not compute the threshold date inside the query helpers, but we should
compute it before and pass the threshold date to query helpers.