✨(api) integrate postgresql database support from the api
We already have a class used by Celery workers to connect to the PostgreSQL
database but it is not ideal for connecting from the API (no connection pool
used, one session per transaction).
Adding a singleton pattern for creating a single engine along with a session
generator, to share a single connection pool throughout the API's lifespan.