🔒️(edx) prevent deletion of entries from protected edx tables
Certain edx tables should not allow deletion.
Rather than depending solely on the ON DELETE RESTRICT constraint in the
database, we handle this with SQLAlchemy within Mork.
A check is added to ensure that, if the user being deleted has associated
entries in these protected tables, an error is raised, preventing the deletion.