🐛(edx) fix mysql seeding script for development
Since the introduction of the edx MySQL models, the seeding script has stopped
working.
We rely on factories to generate fake data into the MySQL development database.
To ensure compatibility with both testing and seeding use cases, the factories
need some proper configuration, mainly on handling the session and the
persistency.
Additionally, subfactories were used to cascade factory creation, which works
in tests but fails when committing to a real development database. These
subfactories have been replaced with factoryboy related factories to resolve the
issue.