🥅(celery) silently exit brevo task for already deleted users
As some users are protected and kept in the edx MySQL database, they can be
scheduled for deletion, on both edX MySQL and Brevo.
At the start of the Brevo task, if user has been deleted, an error is raised.
Instead of raising an error, we should log a warning and silently exit the task,
as the job is done.