🐛(edx) fix student manual enrollment audit relation
Deleting a user from edx MySQL that has is an enroller in the
`student_manualenrollmentaudit` table (field `enrolled_by_id`) returns an
`IntegrityError`.
The relation for field `enrolled_by_id` was missing.
Adding the relation, and since we don't want to delete entries from this table,
and as the `enrolled_by_id` is nullable in the database, we do not specify a
cascade behavior. As a result, the field is set to NULL by SQLAlchemy.